Combine publish and page_maker commands by removing need for explcit page ids.
This change removes the need to manually run the conf_page_maker command by
introducing logic to check for existing pages using the space and page title.
If the page exists, we just store its id, otherwise we create a new page.

We now store the parent page id in the config file, which means we don't need
to pass it via the command line any more.  This is cool, because it allows
you to define more complex page structures in a single config file.